requirements.txt
requests
websockets
loguru
pyyaml
pycryptodome
PyQt5
urllib3==1.26.20
bs4
openpyxl
pykrx
pandas>=2.0.0
lxml>=4.9.0
rapidfuzz>=3.0.0
TA-Lib